+++ src/main/kotlin/com/github/jengelman/gradle/plugins/shadow/internal/DefaultDependencyFilter.kt
@@ -2,9 +2,19 @@ package com.github.jengelman.gradle.plugins.shadow.internal
 
 import com.github.jengelman.gradle.plugins.shadow.tasks.DependencyFilter
 import org.gradle.api.Project
+import org.gradle.api.artifacts.Configuration
 import org.gradle.api.artifacts.ResolvedDependency
+import org.gradle.api.artifacts.component.ModuleComponentIdentifier
+import org.gradle.api.artifacts.component.ProjectComponentIdentifier
+import org.gradle.api.artifacts.result.ResolvedArtifactResult
+import org.gradle.api.artifacts.result.ResolvedDependencyResult
+import org.gradle.api.attributes.Category
+import org.gradle.api.attributes.DocsType
+import org.gradle.api.file.FileCollection
+import org.gradle.jvm.JvmLibrary
+import org.gradle.language.base.artifact.SourcesArtifact
 
-internal class DefaultDependencyFilter(project: Project) :
+internal class DefaultDependencyFilter(@Transient private val project: Project) :
   DependencyFilter.AbstractDependencyFilter(project) {
   override fun resolve(
     dependencies: Set<ResolvedDependency>,
@@ -18,5 +28,80 @@ internal class DefaultDependencyFilter(project: Project) :
         resolve(dep.children, includedDependencies, excludedDependencies)
       }
     }
+  }
+
+  fun resolveSourcesJars(configurations: Collection<Configuration>): FileCollection {
+    return configurations
+      .map { resolveSourcesJars(it) }
+      .reduceOrNull { acc, fileCollection -> acc + fileCollection } ?: project.files()
+  }
+
+  private fun resolveSourcesJars(configuration: Configuration): FileCollection {
+    val includes = mutableSetOf<ResolvedDependency>()
+    val excludes = mutableSetOf<ResolvedDependency>()
+    resolve(
+      dependencies = configuration.resolvedConfiguration.firstLevelModuleDependencies,
+      includedDependencies = includes,
+      excludedDependencies = excludes,
+    )
+
+    val includedDependenciesResults =
+      configuration.incoming.resolutionResult.allDependencies
+        .filterIsInstance<ResolvedDependencyResult>()
+        .filter { dep ->
+          includes.any { inc ->
+            inc.moduleGroup == dep.selected.moduleVersion?.group &&
+              inc.moduleName == dep.selected.moduleVersion?.name &&
+              inc.moduleVersion == dep.selected.moduleVersion?.version
+          }
+        }
+
+    val externalComponentIds =
+      includedDependenciesResults
+        .map { it.selected.id }
+        .filterIsInstance<ModuleComponentIdentifier>()
+        .toSet()
+
+    val externalSourcesFiles =
+      project.dependencies
+        .createArtifactResolutionQuery()
+        .forComponents(externalComponentIds)
+        .withArtifacts(JvmLibrary::class.java, SourcesArtifact::class.java)
+        .execute()
+        .resolvedComponents
+        .flatMap { it.getArtifacts(SourcesArtifact::class.java) }
+        .filterIsInstance<ResolvedArtifactResult>()
+        .map { it.file }
+
+    val projectComponentIds =
+      includedDependenciesResults
+        .map { it.selected.id }
+        .filterIsInstance<ProjectComponentIdentifier>()
+        .toSet()
+
+    val projectSourcesFiles =
+      try {
+        configuration.incoming
+          .artifactView { view ->
+            view.withVariantReselection()
+            view.attributes { attrs ->
+              attrs.attribute(
+                Category.CATEGORY_ATTRIBUTE,
+                project.objects.named(Category::class.java, Category.DOCUMENTATION),
+              )
+              attrs.attribute(
+                DocsType.DOCS_TYPE_ATTRIBUTE,
+                project.objects.named(DocsType::class.java, DocsType.SOURCES),
+              )
+            }
+            view.componentFilter { id -> id in projectComponentIds }
+            view.lenient(true)
+          }
+          .files
+      } catch (_: Exception) {
+        project.files()
+      }
+
+    return project.files(externalSourcesFiles) + projectSourcesFiles
   }
 }